Close realm panel on outside click
